requirements.txt
python-pptx
requests
python-dotenv
